Localized distance units (#324)
Add a toLocalizedDistance() formatter and use that map for multilang-location-standard and multilang-professional-location

The toLocalizedDistance() formatter returns a distance unit of 'mi' or 'km' for the given locale. These mapping are defined in units-18n.js. Any regions not defined in the locale fallback to 'default'. This unit mapping for locales can be expanded to add new types of units.

J=SLAP-623
TEST=manual

Create localized jambo site and observe the different distance units being used on the map for pages with different locales