webpack: Add inline plugin for js/css assets
To support page speed initiatives, we add the ability to inline
webpack-compiled assets. This includes the bundle.js and bundle.css that
are compiled in webpack.

To opt-in to inline assets, add the "data-webpack-inline" attribute to
the script/link element you wish to inline. For scripts, we populate the
script tag with the exact contents of the js file. For links, we replace
the link tag with a style tag with the contents of the css file.

J=SLAP-663
TEST=manual

Test that the expected new case works:
Add to the layouts/html.hbs
        <script src="bundle.js" data-webpack-inline></script>
Add to the core.hbs
        <link rel="stylesheet" type="text/css" href="bundle.css" data-webpack-inline>
npx jambo build && grunt webpack
Make sure the build is successful and content for bundle.js and
bundle.css is not received through a request, but rather inlined in the
DOM.

Test that link/script tags work without data-webpack-inline and don't
try to inline
Test with more than one script/link