Add dns-prefetch for sitescdn and font-display: swap (#333)
dns-prefetching sitescdn, the to perfrom a dns lookup if the browser
has idle time, in theory should only help page speed (outside of
it adding one line of code to the html). Specifying font-display:
swap gives tells the browser it is allowed to use quickly use a
fallback font while loading in a font, and that it has infinite
time to swap in higher priority fonts.

J=SLAP-658
TEST=manual

Used testcafe to run 200 iterations of the page with dns-prefetch
and without dns-prefetch for sitescdn, there was little to no
difference (both averaged around ~800ms startTime for the
yext.answers.verticalQueryResponseRendered performance mark).

Added font-display: swap, and tested that I could still add custom woff
files and specify those are the font.