Get document locale properly (#338)
Remove the 'en-US' locale default so that the openStaus formatter falls back on the document locale

When _getDocumentLocale() was added, it was never being used inside the open status formatter because of the 'en-US' locale default. This PR changes openStatus to fallback on _getDocumentLocale() which lets it support multi-lang sites.

J=None
TEST=manual

View an internationalized jambo site and confirm the English, Spanish, and French translations appear depending on the document locale.