Separate ResultsHeader from VerticalResults (#335)
* Separate ResultsHeader from VerticalResults

This commit deprecates usage of the yxt-ResultsHeader built
into VerticalResults, using the hideResultsHeader config,
and opts to use the new VerticalResultsCount and AppliedFilters
components. This is needed for later collapsible filters PRs.

TEST=manual

These changes were cherry picked off of the collapsible filters work.
Tested that with the full I can load up vertical-grid, standard, and
map pages with the new results header components.