formatters: Split into separate exportables
We split out the formatters into exportable functions in
formatters-internal.js that are combined into one object in
formatters.js. This is to enable HHs to comment out formatters that are
not used in the code. This also decreases the bundle size when a
formatter is commented out. This aligns with our efforts to increase
page speed on answers experiences.

We use this method of tree shaking because Formatters may be used in code
that is not processed by webpack (for example, openStatus formatter
calls in the result card js).

J=SLAP-659
TEST=manual

Test on a new HH theme experience that a search (e.g. "all") works
without console errors (esp. errors refeerencing a missing Formatter).

To test that code is correctly being shaken when unused, try commenting
out a formatter in the Formatters const in formatters.js, e.g.
openStatus
Add a console.log to an openStatus dependency e.g.
```
export function _calculateYextDayTime(now, timezoneUtcOffset) {
  console.log('BOOOOOOOOOOOOOOOOOOO');
```
npx jambo build && grunt webpack
Make sure BOOOOOOOOO is not in the resulting bundle.js