Add open status key param (#342)
Add key param to the openStatus formatter

Add the key param as the second parameter. Move locale from the second param to the last param. Do this to make sure that you can specify isTwentyFourHours without needing to specify the locale, and the hours filed without needing to call out isTwentyFourHours + locale.

The openStatus formatter method signature was approved by Rose.

J=SLAP-630
TEST=manual

Run a jambo build and confirm that the formatter works when we only supply the first profile param. Supply hours data on a different key in the profile and point the formatter to that and confirm it works