Collapsible Filters SCSS (#344)
This commit adds the styling required for collapsible filters.
This includes styling for the two new theme components,
ViewResultsButton and FilterLink, and overrides for SDK filter styling.

One of the requirements was that vertical-standard and vertical-grid
would get collapsible filters, but ONLY on mobile. To do this, I created
a base.scss for collapsible filters styling, that is imported for all
screen widths for vertical-map, but only on mobile for vertical-grid and
standard. CollapsibleFilters also uses the `order` css property
to reorder the results, filters, and resultsheader, so that the results
are on the top, which is necessary to get the results header sticky
at the top but ONLY on mobile. 

This also lets somebody remove the CollapsibleFilters
css class, if they want the old filters layout, and the order of the
3 elements will be filters, then results header, then vertical results.

This commit also tweaks the vertical-map styling, so that now, the Answers-resultsColumn
and Answers-map will still have a min-height of 950px, but will grow to fit
the rest of the page if the page is taller than that. The min-height of 950px
is still necessary so that iframe implementations work properly. To do this
properly, I fixed the height of the Answers-navWrapper to 113px (which
appears to be what the current height is), and set that as a new variable.
This was done because the ViewResultsButton is fixed to the bottom of the page,
but for pages taller than 950 + 113 px, this would look strange, because
it would float in the empty space under the results, instead of at the
bottom of where the results ended.

The view results button also has 2 new variables for controlling its styling
(maybe the padding-top one isn't necessary), one for its height and the
other for its hover color. It doesn't use --yxt-color-hover like you might
expect, because a giant blue button flashing dark grey (which is what
the default for --yxt-color-hover) looked really strange.

TEST=manual

test the below scenarios for grid, standard, and map

with CollapsibleFilters css class
- with facets and Collapsible Filters turned on
- with facets and Collapsible Filters turned on

without CollapsibleFilters css class
- with facets
- without facets

test that the vertical-map still works on an iframe

test that vertical-grid + vertical-standard only get collapsible
filters on mobile

test the sticky results header on mobile for grid + standard, and always
on map. It doesn't work on ie11, because position: sticky doesn't work on ie11,
but the sticky navWrapper (which existed prior to this) also doesn't work on ie11