Fix resultsHeader and filters order
The order css property was removed in #344,
which was being used to visually reorder elements. That PR only
updated the vertical-map page.html.hbs, and not the other 2
page templates.

TEST=manual

See that without this change, on vertical-standard/grid, the results
header is not sticky with CollapsibleFilters on mobile.
See that after swapping the order, the results header is correctly
sticky on mobile with the CollapsibleFilters css class.