Support localized iframe experiences (#349)
Support a data-path attribute on the div with id="answers-container" to support localized iframe experiences

The data-path should be a path to the index file of a localized answers site. For example, if the French experience is located in the "fr" folder of "public" or "desktop", the data-path should be "fr/index.html". If a Spanish site was built with a URL override such as "{pageName}.{locale}.{pageExt}", the data-path should be "index.es.html". If data-path attribute is not specified, the iframe behavior is the same and will load the default experience.

J=SLAP-722
TEST=manual

Run an internationalized jambo site with an iframe test and confirm the data path works for "fr/index.html" and "index.es.html". Test multiple refreshes to confirm the iframe never breaks.