Add truthy check for iconUrl in cards before adding prefix (#354)
As-is, there is an issue where the relativePath is being prepended to the iconUrl in cards even if iconUrl is not specified. Updating to only add relative path prefix if iconUrl is present. As part of this, I checked the other relative path prefix additions in this PR (#329) to ensure that they are null checking before adding the relative path prefix.

TEST=manual
J=none

Build jambo site locally, without an iconUrl specified in the events-standard card before change. See icon not present; inspect DOM and see "iconUrl": "/". Build again after change, see icon present, inspect DOM and see "iconUrl": "".