CollapsibleFilters: make ViewResultsButton sticky to bottom
This commit makes the view results button sticky to the bottom
of the page. It does this by changing Answers-resultsWrapper
to be position: relative, and swapping between position: fixed
and position: absolute for the ViewResultsButton, depending on
whether the user has scrolled beyond Answers-resultsWrapper.

TEST=manual

test that on chrome and ie11, the ViewResultsButton sticks to the
bottom of the page for a vertical-map page with cfilters

test that vertical-grid + standard still work, both without any filters
and with facets + cfilters