Overlay: Add external interface and config classes (#360)
Adds the YextAnswersOverlay class with its init function. Adds OverlayConfig class that applies the defaults and parses the config from the user.

TEST=manual,compile

Test this with the other code in the prototype, see it work as expected. Without the other code in the prototype, run a `grunt webpack` and see that it compiles and an `overlay.js` file is output, as expected.