Set up communication between iFrame and parent frame (#361)
Creates classes to listen to events on the parent frame (ParentFrameObserver), child iframe (IFrameObserver), and update the overlay if there are any relevant interactions from the observers (InteractionDirector). Add placeholder classes/methods for a overlay resizer class and a DOM injector class.

TEST=manual,compile

Test that jambo build && grunt webpack compiles with just this code. Test existing prototype with this code and verify expected behavior.