Add overlay suggestions partial by default to universal (#387)
This partial is harmless when the page is not inside the Overlay (no config from the parent frame Iwill be injected unless the Overlay script is on it). Per Rose, this should be here by default, HH's can remove if they don't want it there -- but like I said, it doesn't hurt anything.

TEST=manual

Add new page using jambo page command, see partial is present. Use this path for overlay and see provided prompts appear.