Add InjectedData data model and prepend domain to experiencePath (#384)
We need to prepend the experience path in the overlay with the injected domains. Used this as an opportunity to centralize that logic for the other iframe integration as well!

TEST=manual

Tested using the overlay with only a path in SGS, inspect the DOM and see the domain prepended to the iframe src.