universalsectiontemplates: Add click analytics to view all link
Previously, the Yext Answers analytics were omitted for the View All
links. Add the event information so Yext Answers analytics are run again
for clicks on the View All..

J=SLAP-720
TEST=manual

Test on standard, grid-two-columns, grid-three-columns: Click on a View
More on a vertical section in the Universal Results pages fires an event
in the Network tab for analytics with the event options.

Also very in DOM that the data-eventoptions and data-eventtype
attributes exist on the view all link <a> tag