Update iFrame src url format (#385)
Add missing '/' between iframe domain and path. Add referrerPageUrl param to iframe's src attribute. We do not support query params in the experiencePath config option, so we do not need to check if any other query params currently exist. Per the PM's, the referrerPageUrl should be the current page, not the referrer of this page.

TEST=manual

Run locally, inspect iframe src using Chrome's devtools. See referrerPageUrl passed as a URL param.