Add config option to control iframe background (#391)
The iframe background color needs to be configurable. A basic integration could include customizing this color. A more advanced user may want to add a repeating background pattern. This is being applied to the iframe element in the corporate page's DOM, so we can't use the CSS variables in the theme.

TEST=manual

Add custom background color, see iframe background color update.