Add JS and SCSS placeholders (#390)
Add JS and SCSS placeholders so that solution templates can be made with custom header and footers

Add new css classes YxtPage, YxtFooter, and YexHeader. Add script placeholders on-bundle-load.js and on-document-load.js. Add custom-modules.js placeholder which allows for custom code to be added to entry.js. Remove the empty div on the universal-standard page.

J=SLAP-728
TEST=manual

Run a build and confirm it works without a header or footer. Add custom header and footer js and scss provided by the hitchhikers and confirm it works with these placeholders.