theme: add financial-professional-location card
Adds the financial-professional-location card, modeled off of the
professional-location card except using name instead of firstName
and lastName, to the hitchhikers theme.

Also changes the alignment for the distance div so that distance
is displayed right-aligned instead of next to a person's name.

J=SLAP-744
TEST=manual
Used the financial-professional-location card in a page and saw that
the card renders information correctly and that distance is
aligned to the right.
Used the professional-location card in a page and saw that distance
is aligned to the right.