verticaltemplates: don't show border line when not present
Before, we would see the border-bottom of the sort options container
when the sort options were not present. Turns out this was more general.
Any component that was not the last component in the filtersWrapper div
would have a border bottom show up if it were not present.

To fix this, we move the border-bottom (and thus bottom padding styling
to the inner div), which only shows up when the component is present).

J=SLAP-753
TEST=manual

Test on a local Jambo site.

Test on a vertical grid and vertical standard
(note this border code  was not in the vertical map layout code).

Make sure you can hit a no results and see no borders on the sort
options and facets components.

Make sure you see borders between the components when the results are
present, though.