Add translations for pretty printing a boolean.
The Theme's formatters include prettyPrintObject, which pretty prints different
kinds of values. For booleans, it was printing them as either 'Yes' or 'No'.
These values had to be localized. Currently, 5 languages are supported for
these two strings. The default is English.

J=SLAP-781
TEST=manual

Used the prettyPrintObject formatter in a card. Used various locales and
boolean values. For each combination, I verified that the correct string was
pretty printed.