cards: Update all product cards to use new/updated fields
The product card needs to use subfields with a different format. Because
the details can have rich text, we stop escaping the details.

J=SLAP-696
TEST=manual

Test on a local repository that a card using the product-standard and
product-prominentimage cards shows all expected data.

For example, if you search "Bouquet of Roses" on the Slapshot test
account and look at the Products vertical, you should see an image, a
price, and a description with the correct rich text formatting.