Solution Template Header/Footer QA Tweaks (#408)
Make some updates to the solution template header/footer item based on QA feedback

Add comments to the script overrides. Remove the yxt-page dir and move the footer, header, and page scss files up a directory. Add Yxt prefixes to the header and footer templates.

J=SLAP-798
TEST=manual

Test the solution template header/footer code by Mcgin again and confirm it still works. Test that the comments do not interfere with the rest of the code.