collapsible-filters: center View All Results button (#412)
Previously, the View All Results button for collapsible filters had more
spacing on the right hand side. Change the button width to be the same
width as the rest of the results container. Also remove an offsetting
margin.

J=SLAP-751
TEST=manual

Test that the View x Results button on a collapsible filter in a
vertical-grid vertical page is centered in the vertical results
container.