Handle relative paths when inlining assets
This commit updates the InlineAssetHtmlPlugin to correctly
handle asset urls that don't point to the base directory.

Additionally, it adds a css post processing step to the plugin,
to handle relative url() attributes in the css. This is needed because,
when a url() is imported in a stylesheet, that url is relative to the stylesheet
itself. However, when you inline that same css, all url()s will
now be relative to the webpage, which may be on a different directory
level than the original stylesheet. This was solved by prepending
url() attributes with the relative path of the page to the bundle.css .

J=SLAP-800
TEST=manual

test that inlining works correctly for pages in the base directory

test that inlining works correctly for pages in directories nested
within the base directory, like en/index.html instead of index.html.

for both, make sure that fonts imported using url() are not 404ing.