Never autofocus when parent frame is using mobile width (#420)
Part of the Overlay QA is that the Overlay's experience iframe's search bar should not autofocus when the parent frame is in mobile. This is more complicated than a normal media query because the experience iframe always thinks it's mobile-width (because the frame is narrow width-wise). This PR executes the media query in the parent frame when the iframe is expanded and passes an isMobile boolean to the experience iframe.

TEST=manual

Open and close the Overlay at desktop width (> 767px), see search input autofocus. Close, then open at mobile screen width. See search input not autofocus.