Overlay: Outside iframe, update to use new button iframe (#409)
Follow up to this PR (#407) which moves the Overlay button out of the experience iframe and into its own. This PR updates the mediator to support interfacing between three frames: the parent, the experience iframe and the button frame. It also updates the IFrameObserver class to work for both of the button and the experience.

TEST=manual

Test locally with PR #407 and on the Krispy Kreme demo site. See Overlay expand, collapse, growTaller and growShorter animations work as expected, and button doesn't bounce or flash.