product-prominentimage: add tag as stock status
We add a new tag element to the prominentimage tag. This should show if
the stock status of the Product entity is present and not "In Stock".
This is a hardcoded value for now, as we get a string from the backend.

The tag should be in the top left corner, only take up 75% of the image
if it is long and should be sourced by two new --hh level css variables.

Test the width changes in IE11.