Pagination: hide padding when pagination not present
Padding used to show up even when there was no pagination content. Move
the padding to the div that is present when pagination content is
present.

J=SLAP-752
TEST=manual

Test two verticals: one with few results (Products, in my case) and one
with many results (query `all` on People). Make sure on former, you do
not see a padding on the `Answers-pagination` class. Make sure on latter
you see the padding on the now-present `yxt-Pagination` div.