Handle relative paths when inlining assets (#419)
This commit updates the InlineAssetHtmlPlugin to correctly
handle asset urls that don't point to the base directory. The
assetsMap provided by webpack was a map to fileNames to
webpack sources, like a bundle.js, but we were keying into it
with the full url of the resource, which could be something like
../bundle.js for localized pages.

Additionally, the inlining plugin now transforms all css url()s to include
the relativePath of the page, if the url() is not an absolute url.
This is needed because, url() attributes in an imported stylesheet are relative
to the stylesheet itself. However, when you write css inline, url()s are relative
to the webpage.

J=SLAP-800
TEST=manual

test with urlFormats:
"a/a/a/a/{locale}/{pageName}.{pageExt}"
"{locale}/{pageName}.{pageExt}"
"{pageName}.{pageExt}"

make sure that fonts imported using url() are not 404ing, and css and js are inlined properly

check that url() attributes that start with #, /, data: , or have a // in them are not transformed