Code styling updates to explicate Overlay cross-frame interactions (#442)
Updated the directory structure and naming to better indicate (1) which frame the various JS scripts are running in, and (2) which constants are shared across frames.
(1) Updated the directory structure to better support shared constants and make it clearer what frame the JS is running in.
(2) Renamed "InteractionTypes" -> "ActionTypes". These "ActionTypes" are shared across frames and are not purely user interactions, some of the constants are actions taken by the frames (e.g. initialization, or onReady events).

TEST=manual

Serve locally, see no console errors and Overlay appears.