Update Overlay config for new integration instructions (#444)
The new integration instructions require setting the configuration on the window rather than calling the YxtAnswersOverlay.init function. Updated the yxtanswersoverlay.js script to initialize itself with the settings from the window. The new integration instructions also allow a user to configure the domain rather than having the injected data handle it - this makes it easier if a developer wants to test the production Overlay before pushing to production.

TEST=manual

Test with new integration instructions, confirm Overlay displays and is responsive to clicks.