custom commands: make all methods except execute static
Changes the `card` and `direactanswercard` commands so that all methods
except `execute` are static.

The commands now export two things: the class itself and a way to
instantiate the class.

J=SLAP-817
TEST=manual
Used the `card` and `directanswercard` commands in a Jambo repo and
saw that there were no behavior changes.