script.hbs: Add JAMBO_INJECTED_DATA fallback for verticalPages label (#453)
* script.hbs: Add JAMBO_INJECTED_DATA fallback for verticalPages label

We added new metadata for the JAMBO_INJECTED_DATA in the Yext CI system.
We now want to use displayName as a default value for the vertical label
in verticalPages if it exists.

Because this lookup is farily large, moved to an inline partial.

Note: Ideally, this logic would not be in the handlebars, and the
templating system would be as dumb as possible. However, we are
restricted by how Jambo injects the environment variables and how we
don't use something like Wepback to do this logic in JavaScript.

J=SLAP-804
TEST=manual

Tested on a variety of page types:
vertical-grid
vertical-map
vertical-standard
universal-standard

Logged out which conditional each vertical met in the `jambo build`. The
universal was always hardcoded (as expected). The `events` and `products`
verticals always fellback to the vertical key (as expected) as the
JAMBO_INJECTED_DATA did not have a displayName for either of those
(empty string and undefined respectively). All other verticals correctly
used the displayName (shown in the Navigation tabs for example) as the
JAMBO_INJECTED_DATA had displayName data for them.

* Change label param to overridedLabel

* Remove logs

* Universal does not need to use the verticalLabel partial, has no verticalKey