webpack: Remove minifying of inline js/css in html
Currently, the Webpack build can be quite slow. We did some profiling
and discovered the source of the problem: the html loader is attempting
to minify all inline JS for each HTML file. This is redundant, however.
Jambo's babel helper already ensures that the inline JS is minified as
part of `jambo build`. Removing minification from the loader should
really speed things up. We tested it out on YAnswers and it halved the
Webpack build time.

Note that html loader is not minifying `bundle.js` (even if it is
inlined). That happens elsewhere and doesn't need to be messed with.

We remove CSS minification as well as there is no inline CSS we
anticipate we are optimizing for and the tradeoffs for speed may not be
worth it if someone were to custom add a bit of inline CSS.

On a local test (Slapshot test account) with 13 configs/pages:
With minification in the theme:
	Command being timed: "grunt webpack"
	User time (seconds): 18.47
	System time (seconds): 1.20

Without minification in the theme:
	Command being timed: "grunt webpack"
	User time (seconds): 11.70
	System time (seconds): 1.01

A small tradeoff to note: The file size of the html files are slightly
bigger with this change.

For example the size of a given HTML file
With minification in the theme: 754K
Without minification in the theme: 844K

J=SLAP-845
TEST=manual

Test on two different accounts.
Test that the experience still runs smoothly and the assets are
reasonably sized