formatters: Change Google Place ID endpoint for get directions (#460)
Currently we use the place endpoint e.g.
https://www.google.com/maps/place/?q=place_id:ChIJQ_9cLMiXyFYRfcmQy7GHHaw

This works great on desktop and mobile if you paste the link directly.
However, if you click a link with this href and have the Google Maps app
on your mobile device, you get a malformed search on the app.

We change the endpoint to be the standard endpoint for get directions
except we add a query parameter for the place id if it exists.

J=SLAP-846
TEST=manual

Tested on a production site by changing the formatters-internal file
locallly and building. Serve the site locally and try to visit a Get
Directions link on a mobile device for a location with a Place ID. This
should open Google Maps with a listings card for the business at the
right location.