cards: Add product-prominentimage-clickable cardType
We add a product card variant that makes the entire card clickable. This
is a copy of the product-prominentimage card. However, a caveat of a
top-level link is that we can no longer have links inside of the card,
otherwise we get invalid HTML. Thus, we remove all options for buttons
and other interactive elements (Show More/Less Button, Title Link,
CTAs).

The url should come from profile.url. It should fallback to
landingPageUrl if this does not exist, and null otherwise.

J=SLAP-629
TEST=manual

Test on both a multilang card and the regular card. (fr in my case)
Use cardType: product-prominentimage-clickable in the config/product.json
Test that in a grid layout and on a regular layout the entire card is
clickable, respects the link target, and sends an analytics event on
click.