sticky view results button in an iframe (#468)
This commit makes the collapsible filters view results button sticky
even when inside an iframe. It does this by using the iframe resizer's
getPageInfo listener, which will be called any time the the parent
frame is scrolled or resized.

An basic css animation was added to make the transition less choppy.

J=SLAP-805
TEST=manual

test with a basic test repo, and then the pizza hut repo,
that the the view results button is sticky both when inside an iframe
and not

tested this works in ie11 for the pizza hut repo. 
it looks like the current develop branch does not work in ie11, likely due to webpack
changes, so the current pizza hut webpack config was used instead