Support ie11 with webpack 5
When migrating for webpack 4 to 5, it looks like
we need to add target: 'es5' to support ie11.
This is suggested on the webpack 4 to 5 migration page
https://webpack.js.org/migrate/5/

From some local testing, it looks like we could get away
with just specifying target: ['es5'] instead of target: ['web', 'es5'].
But it seems more correct to specify 'web' as well, since we're intending
for our code to be run in the browser, and this was also suggested
by the webpack docs.

TEST=manual

test that ie11 works again