image formatter: choose the smallest thumbnail that is big enough
Prior to this commit, the image formatter would always choose the last
thumbnail in the image.thumbnails array.

@creotutar said that the desired behavior is to pick
the smallest one that is at least as big as the `desiredSize`
parameter.

It looks like the image formatter is assuming that the thumbnails
are already sorted largest to smallest, which is not necessarily
the case.

TEST=manual,auto

tested that yanswers now chooses the correct thumbnail
added unit tests