Add 'defer' attribute to SDK Javascript assets. (#473)
This PR adds the 'defer' attribute to the `script` tags for answers.min.js
and answerstemplates.compiled.min.js. This will ensure that the loading of
these JS assets does not block the browser from creating the DOM. This
attribute, as opposed to `async`, allows us to define an ordering between
the `script`s. The templates must be loaded before the SDK itself.

Note that the 'async' attribute was not added to these `script`s. This is
because they need to be executed in the order mentioned above, after the DOM
is built. Further, having both 'async' and 'defer' allows a browser that does
not support 'async' to fallback to 'defer'. We just always want 'defer'.

J=SLAP-860
TEST=manual

Built a local site with these changes. Did some smoke testing and saw no
problems.