use isAbsoluteUrl helper to determine absolute urls (#475)
* use isAbsoluteUrl helper to determine absolute urls

This commit adds the isAbsoluteUrl helper to determine urls,
and updates the regex we were using to also detect urls of the
 mailto:slapshot@gmail.com. In order to use this helper outside of
templates sent to the SDK, this helper also had to be added
to jambo, and the jambo version incremented to 1.9.0.
When we implement custom hbs helpers in jambo, this helper should
be moved fully to the hitchhiker theme.

J=SLAP-861
TEST=manual

test that urls like //yext.com will still be recognized as absolute
test using a data:image/gif;base64, url for favicon (uses jambo helper)
test using a mailto: link inside the professional location card