Update the price formatter and have all Product cards use it. (#476) (#477)
It turns out that LiveAPI returns currencyCode with the display name appended.
As an example, it returns 'USD-US Dollars' instead of just 'USD'. We need to
update our price formatter to account for this. This PR also ensures that the
various Product cards use the new formatter.

J=SLAP-785
TEST=auto, manual

Updated the formatter's unit tests. Created a local site using the various
Product cards and verified that price was formatted properly.