html-asset-loader: use webpack alias and absolute path for static assets
This commit updates the html asset loader to transform static asset
urls into absolute paths, like "/my-img.1234abcd.svg". It also
adds a webpack alias to the static folder, so that the loader can
use the static asset paths regardless of the location of the
html file.

The isAbsoluteUrl helper had to be updated to recognize urls
that only begin with a single slash as absolute urls as well.

J=SLAP-881
T=https://yextops.zendesk.com/agent/tickets/366842
TEST=manual

test that the WHO site can specify an iconUrl that points to a static
asset

tested that a fresh site can also specify an iconUrl in a CTA
that points to a static asset, and also can directly put in
img tags with a static asset src