Fix parentIFrame race condition with sticky view results button
There was a chance of the collapsible filters interactions class
trying to access window.parentIFrame before it was initialized.

This commit fixes that by adding an iframeLoaded promise to the
window, which will be waited on before trying to use the
iframe resizers getPageInfo() listener.

TEST=manual

tested that the scroll listener is never called on an iframe page,
and the resize listener is debounced with the iframe getPageInfo listener

tested on both ie11 and chrome, for both iframed and non-iframed pages,
that the sticky view results button works,

tested that even if I delay the stickifyViewResultsButton() call,
the listeners will be registered correctly