Fix Overlay overlapping issue (#485)
Previously, the elements within the overlay container had a z-index values but the container itself didn't. This was causing issues where the container could be under other items on the page. Add zindex to the container.

TEST=manual
J=PB-9812

Reproduce issue on Krispy Kreme site - see menu labels appear on top of the overlay. Test after change and see overlay is on top of all other elements.