Default JAMBO_INJECTED_DATA (#490)
Default JAMBO_INJECTED_DATA to null

Since Webpack 5, a webpack build will throw an error if an environment variable does not have a value or a default. This causes builds to fail when building locally when JAMBO_INJECTED_DATA is not set. Defaulting to null resolves this issue.

J=none
Test=manual

Build a site locally and confirm that it now builds when JAMBO_INJECTED_DATA is not set