Default universal section title to vertical's `displayName`.
Recently, the `JAMBO_INJECTED_DATA` was updated so that the `displayName` of
each vertical in an experience is provided. We should default to this name
when constructing the vertical's universal section template. As before, if
a `label` or `secionTitleName` has been specified in page configuration, we
will favor those instead.

J=SLAP-887
TEST=manual

Tested the following:

- Set a `label` for a vertical, verified that was used as the section title.
- Set a `sectionTitleName` for a vertical, verified that was used as the
  section title.
- When no `label` or `sectionTitleName` was present, saw the `displayName` from
  the injected data used as the section title.
- If no `diplsayName` was present for the vertical, saw it's `verticalKey` used
  as the section title.